bfsi-ocr-project/requirements.txt
numpy
pandas
pytesseract
pillow
re
opencv-python
pdfplumber
matplotlib
requests
beautifulsoup4
time
flask
scikit-learn
torch
transformers
